refactor: split oversized frontend components (#168)
- DependencyGraph.tsx: extract Legend, Toolbar, NodeDetail, utils (-42%)
- WorkloadsPage.tsx: extract ImportTab, InventoryTab, FormDialog (-61%)
- VersionDiff.tsx: extract InlineView, SideBySide views (-64%)
- WorkloadMapper.tsx: extract SourcePane, TargetPane (-40%)
- 11 new focused sub-components, all 957 tests pass
