+++ frontend/src/components/migration/WorkloadMapper.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useEffect, useState } from "react";
 import {
-  Badge,
   Button,
   MessageBar,
   MessageBarBody,
@@ -11,11 +10,12 @@ import {
 } from "@fluentui/react-components";
 import {
   ArrowSyncRegular,
-  CheckmarkCircleRegular,
   WarningRegular,
 } from "@fluentui/react-icons";
 import type { WorkloadMappingRecord, WorkloadRecord } from "../../services/api";
 import { api } from "../../services/api";
+import WorkloadMapperSourcePane from "./WorkloadMapperSourcePane";
+import WorkloadMapperTargetPane from "./WorkloadMapperTargetPane";
 
 // ---------------------------------------------------------------------------
 // Styles
@@ -44,147 +44,10 @@ const useStyles = makeStyles({
     gridTemplateColumns: "1fr 1fr",
     gap: tokens.spacingHorizontalL,
   },
-  pane: {
-    display: "flex",
-    flexDirection: "column",
-    gap: tokens.spacingVerticalS,
-  },
-  paneTitle: {
-    fontSize: tokens.fontSizeBase400,
-    fontWeight: tokens.fontWeightSemibold,
-    paddingBottom: tokens.spacingVerticalXS,
-    borderBottomWidth: "2px",
-    borderBottomStyle: "solid",
-    borderBottomColor: tokens.colorBrandBackground,
-  },
-  workloadCard: {
-    display: "flex",
-    flexDirection: "column",
-    gap: tokens.spacingVerticalXS,
-    padding: tokens.spacingVerticalS,
-    paddingLeft: tokens.spacingHorizontalS,
-    paddingRight: tokens.spacingHorizontalS,
-    borderRadius: tokens.borderRadiusMedium,
-    borderTopWidth: "1px",
-    borderBottomWidth: "1px",
-    borderLeftWidth: "1px",
-    borderRightWidth: "1px",
-    borderTopStyle: "solid",
-    borderBottomStyle: "solid",
-    borderLeftStyle: "solid",
-    borderRightStyle: "solid",
-    borderTopColor: tokens.colorNeutralStroke1,
-    borderBottomColor: tokens.colorNeutralStroke1,
-    borderLeftColor: tokens.colorNeutralStroke1,
-    borderRightColor: tokens.colorNeutralStroke1,
-    backgroundColor: tokens.colorNeutralBackground1,
-  },
-  workloadCardDragging: {
-    opacity: "0.5",
-    backgroundColor: tokens.colorBrandBackground2,
-  },
-  workloadName: {
-    fontWeight: tokens.fontWeightSemibold,
-    fontSize: tokens.fontSizeBase300,
-  },
-  workloadMeta: {
-    display: "flex",
-    gap: tokens.spacingHorizontalXS,
-    flexWrap: "wrap",
-    alignItems: "center",
-  },
-  mappingRow: {
-    display: "flex",
-    alignItems: "center",
-    gap: tokens.spacingHorizontalXS,
-    marginTop: tokens.spacingVerticalXS,
-  },
-  mappingText: {
-    fontSize: tokens.fontSizeBase200,
-    color: tokens.colorNeutralForeground3,
-  },
-  subscriptionCard: {
-    display: "flex",
-    flexDirection: "column",
-    gap: tokens.spacingVerticalXS,
-    padding: tokens.spacingVerticalS,
-    paddingLeft: tokens.spacingHorizontalS,
-    paddingRight: tokens.spacingHorizontalS,
-    borderRadius: tokens.borderRadiusMedium,
-    borderTopWidth: "2px",
-    borderBottomWidth: "2px",
-    borderLeftWidth: "2px",
-    borderRightWidth: "2px",
-    borderTopStyle: "dashed",
-    borderBottomStyle: "dashed",
-    borderLeftStyle: "dashed",
-    borderRightStyle: "dashed",
-    borderTopColor: tokens.colorNeutralStroke2,
-    borderBottomColor: tokens.colorNeutralStroke2,
-    borderLeftColor: tokens.colorNeutralStroke2,
-    borderRightColor: tokens.colorNeutralStroke2,
-    backgroundColor: tokens.colorNeutralBackground2,
-    minHeight: "64px",
-  },
-  subscriptionCardOver: {
-    borderTopColor: tokens.colorBrandBackground,
-    borderBottomColor: tokens.colorBrandBackground,
-    borderLeftColor: tokens.colorBrandBackground,
-    borderRightColor: tokens.colorBrandBackground,
-    backgroundColor: tokens.colorBrandBackground2,
-  },
-  subscriptionName: {
-    fontWeight: tokens.fontWeightSemibold,
-    fontSize: tokens.fontSizeBase300,
-  },
-  subscriptionPurpose: {
-    fontSize: tokens.fontSizeBase200,
-    color: tokens.colorNeutralForeground3,
-  },
-  assignedList: {
-    display: "flex",
-    flexDirection: "column",
-    gap: tokens.spacingVerticalXS,
-    marginTop: tokens.spacingVerticalXS,
-  },
-  assignedChip: {
-    display: "flex",
-    alignItems: "center",
-    gap: tokens.spacingHorizontalXS,
-    fontSize: tokens.fontSizeBase200,
-    padding: `${tokens.spacingVerticalXXS} ${tokens.spacingHorizontalXS}`,
-    borderRadius: tokens.borderRadiusSmall,
-    backgroundColor: tokens.colorNeutralBackground3,
-  },
   warningsSection: {
     display: "flex",
     flexDirection: "column",
     gap: tokens.spacingVerticalXS,
-  },
-  errorText: {
-    color: tokens.colorPaletteRedForeground1,
-  },
-  emptyHint: {
-    fontSize: tokens.fontSizeBase200,
-    color: tokens.colorNeutralForeground3,
-    fontStyle: "italic",
-  },
-  confidenceHigh: {
-    color: tokens.colorPaletteGreenForeground1,
-  },
-  confidenceMed: {
-    color: tokens.colorPaletteMarigoldForeground1,
-  },
-  confidenceLow: {
-    color: tokens.colorPaletteRedForeground1,
-  },
-  groupLabel: {
-    fontSize: tokens.fontSizeBase200,
-    fontWeight: tokens.fontWeightSemibold,
-    color: tokens.colorNeutralForeground2,
-    textTransform: "uppercase",
-    letterSpacing: "0.05em",
-    marginTop: tokens.spacingVerticalS,
   },
 });
 
@@ -207,12 +70,6 @@ interface WorkloadMapperProps {
 // ---------------------------------------------------------------------------
 // Helpers
 // ---------------------------------------------------------------------------
-
-function confidenceLabel(score: number): string {
-  if (score >= 0.8) return "High";
-  if (score >= 0.5) return "Medium";
-  return "Low";
-}
 
 function groupByType(workloads: WorkloadRecord[]): Record<string, WorkloadRecord[]> {
   const groups: Record<string, WorkloadRecord[]> = {};
@@ -352,7 +209,7 @@ export default function WorkloadMapper({ projectId, subscriptions: propSubscript
       }
       setDraggingId(null);
     },
-    [draggingId, subscriptions, workloads]
+    [draggingId, subscriptions, workloads, overrides, mappings]
   );
 
   // ---------------------------------------------------------------------------
@@ -399,110 +256,25 @@ export default function WorkloadMapper({ projectId, subscriptions: propSubscript
       {/* Split panes */}
       <div className={styles.panes}>
         {/* Left — workload list */}
-        <div className={styles.pane}>
-          <Text className={styles.paneTitle}>Workloads</Text>
-          {workloads.length === 0 && (
-            <Text className={styles.emptyHint}>No workloads found for this project.</Text>
-          )}
-          {Object.entries(groupedWorkloads).map(([type, wls]) => (
-            <div key={type}>
-              <Text className={styles.groupLabel}>{type}</Text>
-              {wls.map((wl) => {
-                const mapping = mappingByWorkloadId[wl.id];
-                const isDragging = draggingId === wl.id;
-                return (
-                  <div
-                    key={wl.id}
-                    className={`${styles.workloadCard}${isDragging ? ` ${styles.workloadCardDragging}` : ""}`}
-                    draggable
-                    onDragStart={() => handleDragStart(wl.id)}
-                    onDragEnd={handleDragEnd}
-                    data-testid={`workload-card-${wl.id}`}
-                  >
-                    <Text className={styles.workloadName}>{wl.name}</Text>
-                    <div className={styles.workloadMeta}>
-                      <Badge appearance="outline" size="small">{wl.criticality}</Badge>
-                      {wl.compliance_requirements.length > 0 && (
-                        <Badge appearance="tint" color="warning" size="small">
-                          {wl.compliance_requirements.join(", ")}
-                        </Badge>
-                      )}
-                    </div>
-                    {mapping && (
-                      <div className={styles.mappingRow}>
-                        <CheckmarkCircleRegular />
-                        <Text className={styles.mappingText}>
-                          {mapping.recommended_subscription_name}
-                        </Text>
-                        <Badge appearance="tint" color="success" size="small">
-                          Suggested
-                        </Badge>
-                        <Text
-                          className={
-                            mapping.confidence_score >= 0.8
-                              ? styles.confidenceHigh
-                              : mapping.confidence_score >= 0.5
-                              ? styles.confidenceMed
-                              : styles.confidenceLow
-                          }
-                        >
-                          {confidenceLabel(mapping.confidence_score)} (
-                          {Math.round(mapping.confidence_score * 100)}%)
-                        </Text>
-                      </div>
-                    )}
-                  </div>
-                );
-              })}
-            </div>
-          ))}
-        </div>
+        <WorkloadMapperSourcePane
+          workloads={workloads}
+          groupedWorkloads={groupedWorkloads}
+          mappingByWorkloadId={mappingByWorkloadId}
+          draggingId={draggingId}
+          onDragStart={handleDragStart}
+          onDragEnd={handleDragEnd}
+        />
 
         {/* Right — subscription targets */}
-        <div className={styles.pane}>
-          <Text className={styles.paneTitle}>Target Subscriptions</Text>
-          {subscriptions.length === 0 && (
-            <Text className={styles.emptyHint}>
-              No subscriptions available. Generate an architecture first.
-            </Text>
-          )}
-          {subscriptions.map((sub) => {
-            const assignedIds = workloadsBySubscription[sub.id] ?? [];
-            const isOver = dragOverSubId === sub.id;
-            return (
-              <div
-                key={sub.id}
-                className={`${styles.subscriptionCard}${isOver ? ` ${styles.subscriptionCardOver}` : ""}`}
-                onDragOver={(e) => {
-                  e.preventDefault();
-                  setDragOverSubId(sub.id);
-                }}
-                onDragLeave={() => setDragOverSubId(null)}
-                onDrop={() => handleDrop(sub.id)}
-                data-testid={`subscription-target-${sub.id}`}
-              >
-                <Text className={styles.subscriptionName}>{sub.name}</Text>
-                <Text className={styles.subscriptionPurpose}>{sub.purpose}</Text>
-                {assignedIds.length > 0 && (
-                  <div className={styles.assignedList}>
-                    {assignedIds.map((wlId) => {
-                      const m = mappingByWorkloadId[wlId];
-                      return (
-                        <div key={wlId} className={styles.assignedChip}>
-                          <CheckmarkCircleRegular fontSize="12px" />
-                          <Text>{m?.workload_name ?? wlId}</Text>
-                        </div>
-                      );
-                    })}
-                  </div>
-                )}
-                {assignedIds.length === 0 && (
-                  <Text className={styles.emptyHint}>Drop workloads here</Text>
-                )}
-              </div>
-            );
-          })}
-        </div>
+        <WorkloadMapperTargetPane
+          subscriptions={subscriptions}
+          workloadsBySubscription={workloadsBySubscription}
+          mappings={mappings}
+          dragOverSubId={dragOverSubId}
+          onDragOver={setDragOverSubId}
+          onDragLeave={() => setDragOverSubId(null)}
+          onDrop={handleDrop}
+        />
       </div>
 
       {/* Validation warnings */}
