+++ frontend/src/pages/WorkloadsPage.tsx
@@ -1,7 +1,6 @@
-import { useCallback, useRef, useState } from "react";
+import { useCallback, useState } from "react";
 import { useParams } from "react-router-dom";
 import {
-  Badge,
   Button,
   Dialog,
   DialogActions,
@@ -10,45 +9,32 @@ import {
   DialogSurface,
   DialogTitle,
   DialogTrigger,
-  Divider,
-  Field,
-  Input,
   Label,
   MessageBar,
   MessageBarBody,
-  Select,
   Spinner,
   Tab,
   TabList,
-  Table,
-  TableBody,
-  TableCell,
-  TableHeader,
-  TableHeaderCell,
-  TableRow,
   Text,
-  Textarea,
   makeStyles,
   tokens,
 } from "@fluentui/react-components";
 import {
   ArrowUploadRegular,
-  CheckmarkCircleRegular,
-  DeleteRegular,
   DocumentRegular,
-  EditRegular,
   LinkRegular,
   ShareRegular,
-  WarningRegular,
 } from "@fluentui/react-icons";
 import type {
   WorkloadCreateRequest,
-  WorkloadImportResult,
   WorkloadRecord,
 } from "../services/api";
 import { api } from "../services/api";
 import WorkloadMapper from "../components/migration/WorkloadMapper";
 import DependencyGraph from "../components/migration/DependencyGraph";
+import WorkloadImportTab from "./WorkloadImportTab";
+import WorkloadInventoryTab from "./WorkloadInventoryTab";
+import WorkloadFormDialog from "./WorkloadFormDialog";
 
 const useStyles = makeStyles({
   root: {
@@ -69,86 +55,6 @@ const useStyles = makeStyles({
     flexDirection: "column",
     gap: tokens.spacingVerticalM,
   },
-  dropZone: {
-    border: `2px dashed ${tokens.colorNeutralStroke1}`,
-    borderRadius: tokens.borderRadiusMedium,
-    padding: tokens.spacingVerticalXXL,
-    textAlign: "center",
-    cursor: "pointer",
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    gap: tokens.spacingVerticalS,
-    ":hover": {
-      backgroundColor: tokens.colorNeutralBackground2,
-    },
-  },
-  dropZoneActive: {
-    backgroundColor: tokens.colorBrandBackground2,
-  },
-  previewSection: {
-    display: "flex",
-    flexDirection: "column",
-    gap: tokens.spacingVerticalS,
-  },
-  errorList: {
-    display: "flex",
-    flexDirection: "column",
-    gap: tokens.spacingVerticalXS,
-  },
-  tableWrapper: {
-    overflowX: "auto",
-  },
-  actionRow: {
-    display: "flex",
-    gap: tokens.spacingHorizontalM,
-    alignItems: "center",
-    justifyContent: "space-between",
-  },
-  formGrid: {
-    display: "grid",
-    gridTemplateColumns: "1fr 1fr",
-    gap: tokens.spacingVerticalM,
-  },
-  formFullRow: {
-    gridColumn: "1 / -1",
-  },
-  importSummary: {
-    display: "flex",
-    gap: tokens.spacingHorizontalM,
-    alignItems: "center",
-  },
-  uploadIcon: {
-    fontSize: "32px",
-    color: tokens.colorBrandForeground1,
-  },
-  browseHint: {
-    color: tokens.colorNeutralForeground3,
-  },
-  hiddenInput: {
-    display: "none",
-  },
-  buttonRow: {
-    display: "flex",
-    gap: tokens.spacingHorizontalS,
-  },
-  successIcon: {
-    fontSize: "20px",
-    color: tokens.colorStatusSuccessForeground1,
-  },
-  failedCount: {
-    color: tokens.colorStatusWarningForeground1,
-  },
-  warningText: {
-    color: tokens.colorStatusWarningForeground1,
-  },
-  actionCellButtons: {
-    display: "flex",
-    gap: tokens.spacingHorizontalXS,
-  },
-  saveErrorBar: {
-    marginTop: tokens.spacingVerticalM,
-  },
   deleteButton: {
     backgroundColor: tokens.colorStatusDangerBackground3,
   },
@@ -158,11 +64,6 @@ const useStyles = makeStyles({
 });
 
 type TabValue = "import" | "inventory" | "mapping" | "dependencies";
-
-const WORKLOAD_TYPES = ["vm", "database", "web-app", "container", "other"];
-const SOURCE_PLATFORMS = ["vmware", "hyperv", "physical", "aws", "gcp", "other"];
-const CRITICALITY_OPTIONS = ["mission-critical", "business-critical", "standard", "dev-test"];
-const MIGRATION_STRATEGIES = ["rehost", "refactor", "rearchitect", "rebuild", "replace", "unknown"];
 
 interface WorkloadsPageProps {
   projectId?: string;
@@ -193,14 +94,6 @@ export default function WorkloadsPage({ projectId: propProjectId }: WorkloadsPag
   const styles = useStyles();
   const [activeTab, setActiveTab] = useState<TabValue>("import");
 
-  // Import tab
-  const [dragOver, setDragOver] = useState(false);
-  const [importFile, setImportFile] = useState<File | null>(null);
-  const [importResult, setImportResult] = useState<WorkloadImportResult | null>(null);
-  const [importLoading, setImportLoading] = useState(false);
-  const [importError, setImportError] = useState<string | null>(null);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
   // Inventory tab
   const [workloads, setWorkloads] = useState<WorkloadRecord[]>([]);
   const [listLoading, setListLoading] = useState(false);
@@ -217,43 +110,6 @@ export default function WorkloadsPage({ projectId: propProjectId }: WorkloadsPag
   const [deleteTarget, setDeleteTarget] = useState<WorkloadRecord | null>(null);
   const [deleteLoading, setDeleteLoading] = useState(false);
   const [deleteError, setDeleteError] = useState<string | null>(null);
-
-  const handleDrop = useCallback((e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    setDragOver(false);
-    const file = e.dataTransfer.files[0];
-    if (file) {
-      setImportFile(file);
-      setImportResult(null);
-      setImportError(null);
-    }
-  }, []);
-
-  const handleFileSelect = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0] ?? null;
-    if (file) {
-      setImportFile(file);
-      setImportResult(null);
-      setImportError(null);
-    }
-  };
-
-  const handleImport = async () => {
-    if (!importFile) return;
-    setImportLoading(true);
-    setImportError(null);
-    try {
-      const result = await api.workloads.importFile(importFile, projectId);
-      setImportResult(result);
-      if (result.imported_count > 0) {
-        fetchWorkloads();
-      }
-    } catch (err) {
-      setImportError(err instanceof Error ? err.message : "Import failed");
-    } finally {
-      setImportLoading(false);
-    }
-  };
 
   const fetchWorkloads = useCallback(async () => {
     setListLoading(true);
@@ -357,244 +213,23 @@ export default function WorkloadsPage({ projectId: propProjectId }: WorkloadsPag
 
       {/* ---- Import Tab ---- */}
       {activeTab === "import" && (
-        <div className={styles.tabContent}>
-          <Text>Upload a CSV or JSON file to bulk-import workloads into this project.</Text>
-
-          <div
-            className={`${styles.dropZone} ${dragOver ? styles.dropZoneActive : ""}`}
-            onDragOver={(e) => { e.preventDefault(); setDragOver(true); }}
-            onDragLeave={() => setDragOver(false)}
-            onDrop={handleDrop}
-            onClick={() => fileInputRef.current?.click()}
-            role="button"
-            tabIndex={0}
-            onKeyDown={(e) => e.key === "Enter" && fileInputRef.current?.click()}
-            aria-label="Drop zone for workload import file"
-          >
-            <ArrowUploadRegular className={styles.uploadIcon} />
-            <Text weight="semibold">
-              {importFile ? importFile.name : "Drop a CSV or JSON file here"}
-            </Text>
-            <Text size={200} className={styles.browseHint}>
-              or click to browse
-            </Text>
-            <input
-              ref={fileInputRef}
-              type="file"
-              accept=".csv,.json"
-              className={styles.hiddenInput}
-              onChange={handleFileSelect}
-              aria-label="File input for workload import"
-            />
-          </div>
-
-          {importFile && (
-            <div className={styles.actionRow}>
-              <Text>
-                Selected: <strong>{importFile.name}</strong>{" "}
-                ({(importFile.size / 1024).toFixed(1)} KB)
-              </Text>
-              <div className={styles.buttonRow}>
-                <Button
-                  appearance="subtle"
-                  onClick={() => { setImportFile(null); setImportResult(null); setImportError(null); }}
-                >
-                  Clear
-                </Button>
-                <Button
-                  appearance="primary"
-                  onClick={handleImport}
-                  disabled={importLoading}
-                  icon={importLoading ? <Spinner size="tiny" /> : undefined}
-                >
-                  {importLoading ? "Importing\u2026" : "Import"}
-                </Button>
-              </div>
-            </div>
-          )}
-
-          {importError && (
-            <MessageBar intent="error">
-              <MessageBarBody>{importError}</MessageBarBody>
-            </MessageBar>
-          )}
-
-          {importResult && (
-            <div className={styles.previewSection}>
-              <div className={styles.importSummary}>
-                <CheckmarkCircleRegular className={styles.successIcon} />
-                <Text>
-                  Imported{" "}
-                  <strong>{importResult.imported_count}</strong>{" "}
-                  workload{importResult.imported_count !== 1 ? "s" : ""}
-                  {importResult.failed_count > 0 && (
-                    <span className={styles.failedCount}>
-                      {" "}({importResult.failed_count} failed)
-                    </span>
-                  )}
-                </Text>
-              </div>
-
-              {importResult.errors.length > 0 && (
-                <div className={styles.errorList}>
-                  <Text weight="semibold" className={styles.warningText}>
-                    <WarningRegular /> Row errors:
-                  </Text>
-                  {importResult.errors.map((err, i) => (
-                    <Text key={i} size={200} className={styles.warningText}>
-                      {err}
-                    </Text>
-                  ))}
-                </div>
-              )}
-
-              {importResult.workloads.length > 0 && (
-                <>
-                  <Divider />
-                  <Text weight="semibold">Preview ({importResult.workloads.length} rows)</Text>
-                  <div className={styles.tableWrapper}>
-                    <Table>
-                      <TableHeader>
-                        <TableRow>
-                          <TableHeaderCell>Name</TableHeaderCell>
-                          <TableHeaderCell>Type</TableHeaderCell>
-                          <TableHeaderCell>Platform</TableHeaderCell>
-                          <TableHeaderCell>CPUs</TableHeaderCell>
-                          <TableHeaderCell>RAM (GB)</TableHeaderCell>
-                          <TableHeaderCell>Criticality</TableHeaderCell>
-                          <TableHeaderCell>Strategy</TableHeaderCell>
-                        </TableRow>
-                      </TableHeader>
-                      <TableBody>
-                        {importResult.workloads.map((wl) => (
-                          <TableRow key={wl.id}>
-                            <TableCell>{wl.name}</TableCell>
-                            <TableCell>{wl.type}</TableCell>
-                            <TableCell>{wl.source_platform}</TableCell>
-                            <TableCell>{wl.cpu_cores ?? "\u2014"}</TableCell>
-                            <TableCell>{wl.memory_gb ?? "\u2014"}</TableCell>
-                            <TableCell>
-                              <Badge
-                                appearance="tint"
-                                color={
-                                  wl.criticality === "mission-critical"
-                                    ? "danger"
-                                    : wl.criticality === "business-critical"
-                                      ? "warning"
-                                      : "informative"
-                                }
-                              >
-                                {wl.criticality}
-                              </Badge>
-                            </TableCell>
-                            <TableCell>{wl.migration_strategy}</TableCell>
-                          </TableRow>
-                        ))}
-                      </TableBody>
-                    </Table>
-                  </div>
-                </>
-              )}
-            </div>
-          )}
-        </div>
+        <WorkloadImportTab
+          projectId={projectId}
+          onImportSuccess={fetchWorkloads}
+        />
       )}
 
       {/* ---- Inventory Tab ---- */}
       {activeTab === "inventory" && (
-        <div className={styles.tabContent}>
-          <div className={styles.actionRow}>
-            <Text>
-              {workloads.length} workload{workloads.length !== 1 ? "s" : ""} in this project
-            </Text>
-            <div className={styles.buttonRow}>
-              <Button appearance="subtle" onClick={fetchWorkloads}disabled={listLoading}>
-                {listLoading ? <Spinner size="tiny" /> : "Refresh"}
-              </Button>
-              <Button appearance="primary" onClick={openCreate}>
-                Add Workload
-              </Button>
-            </div>
-          </div>
-
-          {listError && (
-            <MessageBar intent="error">
-              <MessageBarBody>{listError}</MessageBarBody>
-            </MessageBar>
-          )}
-          {listLoading && <Spinner label="Loading workloads\u2026" />}
-          {!listLoading && workloads.length === 0 && (
-            <MessageBar intent="info">
-              <MessageBarBody>
-                No workloads yet. Import a file or add one manually.
-              </MessageBarBody>
-            </MessageBar>
-          )}
-          {!listLoading && workloads.length > 0 && (
-            <div className={styles.tableWrapper}>
-              <Table>
-                <TableHeader>
-                  <TableRow>
-                    <TableHeaderCell>Name</TableHeaderCell>
-                    <TableHeaderCell>Type</TableHeaderCell>
-                    <TableHeaderCell>Platform</TableHeaderCell>
-                    <TableHeaderCell>CPUs</TableHeaderCell>
-                    <TableHeaderCell>RAM (GB)</TableHeaderCell>
-                    <TableHeaderCell>OS</TableHeaderCell>
-                    <TableHeaderCell>Criticality</TableHeaderCell>
-                    <TableHeaderCell>Strategy</TableHeaderCell>
-                    <TableHeaderCell>Actions</TableHeaderCell>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {workloads.map((wl) => (
-                    <TableRow key={wl.id}>
-                      <TableCell>{wl.name}</TableCell>
-                      <TableCell>{wl.type}</TableCell>
-                      <TableCell>{wl.source_platform}</TableCell>
-                      <TableCell>{wl.cpu_cores ?? "\u2014"}</TableCell>
-                      <TableCell>{wl.memory_gb ?? "\u2014"}</TableCell>
-                      <TableCell>{wl.os_type ?? "\u2014"}</TableCell>
-                      <TableCell>
-                        <Badge
-                          appearance="tint"
-                          color={
-                            wl.criticality === "mission-critical"
-                              ? "danger"
-                              : wl.criticality === "business-critical"
-                                ? "warning"
-                                : "informative"
-                          }
-                        >
-                          {wl.criticality}
-                        </Badge>
-                      </TableCell>
-                      <TableCell>{wl.migration_strategy}</TableCell>
-                      <TableCell>
-                        <div className={styles.actionCellButtons}>
-                          <Button
-                            appearance="subtle"
-                            size="small"
-                            icon={<EditRegular />}
-                            onClick={() => openEdit(wl)}
-                            aria-label={`Edit ${wl.name}`}
-                          />
-                          <Button
-                            appearance="subtle"
-                            size="small"
-                            icon={<DeleteRegular />}
-                            onClick={() => setDeleteTarget(wl)}
-                            aria-label={`Delete ${wl.name}`}
-                          />
-                        </div>
-                      </TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            </div>
-          )}
-        </div>
+        <WorkloadInventoryTab
+          workloads={workloads}
+          listLoading={listLoading}
+          listError={listError}
+          onRefresh={fetchWorkloads}
+          onOpenCreate={openCreate}
+          onOpenEdit={openEdit}
+          onDeleteTarget={setDeleteTarget}
+        />
       )}
 
       {/* ---- Dependencies Tab ---- */}
@@ -605,144 +240,16 @@ export default function WorkloadsPage({ projectId: propProjectId }: WorkloadsPag
       )}
 
       {/* ---- Create/Edit Dialog ---- */}
-      <Dialog open={dialogOpen} onOpenChange={(_, d) => setDialogOpen(d.open)}>
-        <DialogSurface>
-          <DialogBody>
-            <DialogTitle>{editTarget ? "Edit Workload" : "Add Workload"}</DialogTitle>
-            <DialogContent>
-              <div className={styles.formGrid}>
-                <Field label="Name" required className={styles.formFullRow}>
-                  <Input
-                    value={draft.name}
-                    onChange={(_, d) => setDraft((p: WorkloadCreateRequest) => ({ ...p, name: d.value }))}
-                    placeholder="e.g. web-server-01"
-                  />
-                </Field>
-                <Field label="Type">
-                  <Select
-                    value={draft.type ?? "other"}
-                    onChange={(_, d) => setDraft((p: WorkloadCreateRequest) => ({ ...p, type: d.value }))}
-                  >
-                    {WORKLOAD_TYPES.map((t) => <option key={t} value={t}>{t}</option>)}
-                  </Select>
-                </Field>
-                <Field label="Source Platform">
-                  <Select
-                    value={draft.source_platform ?? "other"}
-                    onChange={(_, d) => setDraft((p: WorkloadCreateRequest) => ({ ...p, source_platform: d.value }))}
-                  >
-                    {SOURCE_PLATFORMS.map((p) => <option key={p} value={p}>{p}</option>)}
-                  </Select>
-                </Field>
-                <Field label="CPU Cores">
-                  <Input
-                    type="number"
-                    value={draft.cpu_cores != null ? String(draft.cpu_cores) : ""}
-                    onChange={(_, d) =>
-                      setDraft((p: WorkloadCreateRequest) => ({ ...p, cpu_cores: d.value ? parseInt(d.value, 10) : null }))
-                    }
-                    placeholder="e.g. 4"
-                  />
-                </Field>
-                <Field label="Memory (GB)">
-                  <Input
-                    type="number"
-                    value={draft.memory_gb != null ? String(draft.memory_gb) : ""}
-                    onChange={(_, d) =>
-                      setDraft((p: WorkloadCreateRequest) => ({ ...p, memory_gb: d.value ? parseFloat(d.value) : null }))
-                    }
-                    placeholder="e.g. 16"
-                  />
-                </Field>
-                <Field label="Storage (GB)">
-                  <Input
-                    type="number"
-                    value={draft.storage_gb != null ? String(draft.storage_gb) : ""}
-                    onChange={(_, d) =>
-                      setDraft((p: WorkloadCreateRequest) => ({ ...p, storage_gb: d.value ? parseFloat(d.value) : null }))
-                    }
-                    placeholder="e.g. 500"
-                  />
-                </Field>
-                <Field label="OS Type">
-                  <Input
-                    value={draft.os_type ?? ""}
-                    onChange={(_, d) => setDraft((p: WorkloadCreateRequest) => ({ ...p, os_type: d.value || null }))}
-                    placeholder="e.g. Windows"
-                  />
-                </Field>
-                <Field label="OS Version">
-                  <Input
-                    value={draft.os_version ?? ""}
-                    onChange={(_, d) => setDraft((p: WorkloadCreateRequest) => ({ ...p, os_version: d.value || null }))}
-                    placeholder="e.g. Server 2022"
-                  />
-                </Field>
-                <Field label="Criticality">
-                  <Select
-                    value={draft.criticality ?? "standard"}
-                    onChange={(_, d) => setDraft((p: WorkloadCreateRequest) => ({ ...p, criticality: d.value }))}
-                  >
-                    {CRITICALITY_OPTIONS.map((c) => <option key={c} value={c}>{c}</option>)}
-                  </Select>
-                </Field>
-                <Field label="Migration Strategy">
-                  <Select
-                    value={draft.migration_strategy ?? "unknown"}
-                    onChange={(_, d) => setDraft((p: WorkloadCreateRequest) => ({ ...p, migration_strategy: d.value }))}
-                  >
-                    {MIGRATION_STRATEGIES.map((s) => <option key={s} value={s}>{s}</option>)}
-                  </Select>
-                </Field>
-                <Field
-                  label="Compliance Requirements"
-                  hint="Comma-separated list of framework IDs"
-                  className={styles.formFullRow}
-                >
-                  <Input
-                    value={(draft.compliance_requirements ?? []).join(", ")}
-                    onChange={(_, d) =>
-                      setDraft((p: WorkloadCreateRequest) => ({
-                        ...p,
-                        compliance_requirements: d.value
-                          ? d.value.split(",").map((s) => s.trim()).filter(Boolean)
-                          : [],
-                      }))
-                    }
-                    placeholder="e.g. SOC2, ISO27001"
-                  />
-                </Field>
-                <Field label="Notes" className={styles.formFullRow}>
-                  <Textarea
-                    value={draft.notes ?? ""}
-                    onChange={(_, d) => setDraft((p: WorkloadCreateRequest) => ({ ...p, notes: d.value || null }))}
-                    placeholder="Optional notes about this workload"
-                    rows={3}
-                  />
-                </Field>
-              </div>
-              {saveError && (
-                <MessageBar intent="error" className={styles.saveErrorBar}>
-                  <MessageBarBody>{saveError}</MessageBarBody>
-                </MessageBar>
-              )}
-            </DialogContent>
-            <DialogActions>
-              <DialogTrigger disableButtonEnhancement>
-                <Button appearance="secondary">Cancel</Button>
-              </DialogTrigger>
-              <Button
-                appearance="primary"
-                onClick={handleSave}
-                disabled={saveLoading}
-                icon={saveLoading ? <Spinner size="tiny" /> : undefined}
-              >
-                {saveLoading ? "Saving\u2026" : editTarget ? "Save Changes" : "Create Workload"}
-              </Button>
-            </DialogActions>
-          </DialogBody>
-        </DialogSurface>
-      </Dialog>
+      <WorkloadFormDialog
+        open={dialogOpen}
+        onOpenChange={setDialogOpen}
+        editTarget={editTarget}
+        draft={draft}
+        onDraftChange={setDraft}
+        saveLoading={saveLoading}
+        saveError={saveError}
+        onSave={handleSave}
+      />
 
       {/* ---- Delete Confirm Dialog ---- */}
       <Dialog open={!!deleteTarget} onOpenChange={(_, d) => { if (!d.open) { setDeleteTarget(null); setDeleteError(null); } }}>
@@ -787,3 +294,4 @@ export default function WorkloadsPage({ projectId: propProjectId }: WorkloadsPag
     </div>
   );
 }
+
